requirements.txt
streamlit
pandas
pillow
python-dotenv
requests